Improved code
Improved the program's code almost everywhere

Added an easter egg (Rename WoRCP.exe to RGBWoRCP.exe)
Added global theme settings
Added the ability to change theme on the fly (Without restarting the app)
Fixed disabled toggles having accent color instead of being gray
Added exception handling
Added more documentation (Phase 2)
Improved config system
Rewrote a lot of old code
Updated settings tab with better code (Phase 2), Still needs a lot of work (Needs to be split apart again after the optimization)
Fixed GPIO tab crashing on non arm devices
Replaced switch case code in theme system with if else (Temporary until a better solution arrives, Necessary for on the fly theme switching)
Removed unused code
Automated the removal of .xml files on build
And many more changes

+++ Tabs/Overlay.cs
@@ -202,7 +202,7 @@ namespace WoRCP.Tabs
         //Opacity slider
         private void OpacitySlider_Selecting(object sender, EventArgs e)
         {
-            OpacityState.Text = ((byte)OpacitySlider.ValueP).ToString() + "%";
+            OpacityState.Text = OpacitySlider.ValueP.ToString() + "%";
             Configuration.overlay.Opacity = OpacitySlider.Value / 100;
         }
         #endregion
